Fixup! added Id implementation responsible for computing _id

+++ testing/src/main/java/io/crate/test/integration/CrateTestCluster.java
@@ -574,7 +574,9 @@ public class CrateTestCluster implements Iterable<Client> {
                 this.dataDirToClean.clear();
             }
         }
-        FileSystemUtils.deleteRecursively(tmpDataDir.toAbsolutePath().toFile());
+        if (tmpDataDir != null) {
+            FileSystemUtils.deleteRecursively(tmpDataDir.toAbsolutePath().toFile());
+        }
     }
 
     /**
